Extract compute level helper functions to a separate crate
