+++ crates/compute/Cargo.toml
@@ -18,5 +18,6 @@ thiserror.workspace = true
 
 [dev-dependencies]
 assert_matches.workspace = true
+binius_compute_test_utils = { path = "../compute_test_utils", default-features = false }
 binius_core = { path = "../core", default-features = false }
 rand = { workspace = true, features = ["std"] }
