+++ crates/compute_test_utils/Cargo.toml
@@ -1,0 +1,19 @@
+[package]
+name = "binius_compute_test_utils"
+version.workspace = true
+edition.workspace = true
+authors.workspace = true
+
+
+[lints]
+workspace = true
+
+[dependencies]
+binius_compute = { path = "../compute", default-features = false }
+binius_core = { path = "../core", default-features = false }
+binius_field = { path = "../field", default-features = false }
+binius_math = { path = "../math", default-features = false }
+binius_ntt = { path = "../ntt", default-features = false }
+binius_utils = { path = "../utils", default-features = false }
+bytemuck = { workspace = true, features = ["extern_crate_alloc"] }
+rand = { workspace = true, features = ["std"] }
