+++ crates/compute_test_utils/src/lib.rs
@@ -1,0 +1,3 @@
+// Copyright 2025 Irreducible Inc.
+
+pub mod layer;
